Move ToolExecutionDeniedError to tool-confirmation module
- Relocate error class from src/core/errors.ts to src/core/client/tool-confirmation/errors.ts
- Update import paths in vercel.ts and manager.ts to reflect new location
- Better organization by keeping tool confirmation errors with related code
