+++ src/core/client/tool-confirmation/errors.ts
